requirements.txt
asgiref==3.9.2
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
cloudscraper==1.2.71
diff-match-patch==20241021
Django==5.2.6
django-import-export==4.3.10
et_xmlfile==2.0.0
fontawesomefree==6.6.0
idna==3.10
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
